[AD-543] bugfix: change data type
 update dataType to int16_t to match DocDb Jdbc Type constnats in binary_common.h

+++ src/odbc/include/ignite/odbc/meta/column_meta.h
@@ -238,7 +238,7 @@ class ColumnMeta {
   std::string remarks;
 
   /** Data type. */
-  int8_t dataType;
+  int16_t dataType;
 
   /** Column precision. */
   int32_t precision;
